+++ lib/awful/init.lua
@@ -51,6 +51,7 @@ return
     remote = require("awful.remote");
     key = require("awful.key");
     button = require("awful.button");
+    wibar = require("awful.wibar");
     wibox = require("awful.wibox");
     startup_notification = require("awful.startup_notification");
     tooltip = require("awful.tooltip");
